pyproject.toml
[project]
name = "dha-quality-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "contourpy==1.3.1",
    "matplotlib==3.10.1",
    "numpy==2.2.5",
    "opencv-python==4.11.0.86",
    "pillow==11.2.1",
    "prefetch-generator==1.0.3",
    "scikit-image==0.25.2",
    "scipy==1.15.2",
    "tensorboardx==2.6.2.2",
    "torch==2.6.0",
    "torchvision==0.21.0",
]
